[MODEXPW-529] Finish CsvMapper test, todo 2 field extractions
