+++ src/main/java/org/folio/dew/batch/acquisitions/edifact/mapper/converter/ClaimCsvFields.java
@@ -18,7 +18,7 @@ public enum ClaimCsvFields implements ExtractableField<Pair<CompositePoLine, Pie
   POL_NUMBER("POL number", (pair) -> pair.getKey().getPoLineNumber()),
   ORDER_NUMBER("Vendor order number", (pair) -> getVendorOrderNumber(pair.getKey())),
   ACCOUNT_NUMBER("Account number", (pair) -> getVendorAccountNumber(pair.getKey())),
-  TITLE("Title", (pair) -> pair.getValue().getTitleId()),
+  TITLE("Title from piece", (pair) -> "CHANGEME"), // FIXME: piece.titleId | poLine.titleOrPackage | title from inventory by piece.titleId
   DISPLAY_SUMMARY("Display summary", (pair) -> pair.getValue().getDisplaySummary()),
   CHRONOLOGY("Chronology", (pair) -> pair.getValue().getChronology()),
   ENUMERATION("Enumeration", (pair) -> pair.getValue().getEnumeration()),
