+++ src/test/resources/edifact/acquisitions/csv_claims_result.csv
@@ -1,0 +1,8 @@
+POL number,Vendor order number,Account number,Title from piece,Display summary,Chronology,Enumeration,Quantity
+10000-1,ORD1000,BRXXXXX-01,CHANGEME,S1,C1,E1,CHANGEME
+10000-1,ORD1000,BRXXXXX-01,CHANGEME,S2,C2,,CHANGEME
+10001-1,ORD1001,BRXXXXX-01,CHANGEME,S3,,E3,CHANGEME
+10001-2,ORD1002,BRXXXXX-01,CHANGEME,,C4,E4,CHANGEME
+10002-1,,,CHANGEME,S5,,E5,CHANGEME
+10002-1,,,CHANGEME,,C5,,CHANGEME
+10002-1,,,CHANGEME,,,,CHANGEME
